+++ src/lib.rs
@@ -29,7 +29,7 @@
 //!
 //! // Create a circuit breaker which configured by reasonable default backoff and
 //! // failure accrual policy.
-//! let circuit_breaker = CircuitBreaker::builder().build();
+//! let circuit_breaker = CircuitBreaker::default();
 //!
 //! // Call the function in a loop, after some iterations the circuit breaker will
 //! // be in a open state and reject next calls.