+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 ### UNRELEASED
 
 Breaking changes:
-
 * `success_rate` policy now accepts `min_request_threshold`.
 
 Improvements:
-
+* add `CircuitBreaker::default()`
 * remove `tokio-timer` dependency.